refactor(demo): Extract KnightTour and TourType to own source file
Move KnightTour class and TourType enum (now nested inside KnightTour)
out of WarnsdorffRuleKnightTourHeuristic.java into KnightTour.java to
resolve auxiliary-class-access and module-accessibility warnings.
